+++ R/read_log.R
@@ -10,10 +10,13 @@
 #' read_log(readr_example("example.log"))
 read_log <- function(file, col_names = FALSE, col_types = NULL,
                      trim_ws = TRUE,
-                     skip = 0, n_max = Inf, progress = show_progress()) {
+                     skip = 0, n_max = Inf,
+                     show_col_types = should_show_types(),
+                     progress = show_progress()) {
   tokenizer <- tokenizer_log(trim_ws = trim_ws)
   read_delimited(file, tokenizer,
     col_names = col_names, col_types = col_types,
-    skip = skip, n_max = n_max, progress = progress
+    skip = skip, n_max = n_max, progress = progress,
+    show_col_types = show_col_types
   )
 }