+++ R/read_table.R
@@ -34,6 +34,7 @@ read_table <- function(file, col_names = TRUE, col_types = NULL,
                         locale = default_locale(), na = "NA", skip = 0,
                         n_max = Inf, guess_max = min(n_max, 1000),
                         progress = show_progress(), comment = "",
+                        show_col_types = should_show_types(),
                         skip_empty_rows = TRUE) {
   tokenizer <- tokenizer_ws(
     na = na, comment = comment,
@@ -42,7 +43,8 @@ read_table <- function(file, col_names = TRUE, col_types = NULL,
   read_delimited(file, tokenizer,
     col_names = col_names, col_types = col_types,
     locale = locale, skip = skip, skip_empty_rows = skip_empty_rows,
-    skip_quote = FALSE, comment = comment, n_max = n_max, guess_max = guess_max, progress = progress
+    skip_quote = FALSE, comment = comment, n_max = n_max, guess_max = guess_max, progress = progress,
+    show_col_types = show_col_types
   )
 }